update keypoint module. need to add tutorials
