+++ gluoncv/model_zoo/simple_pose/simple_pose_resnet.py
@@ -16,6 +16,7 @@ __all__ = ['get_simple_pose_resnet', 'SimplePoseResNet',
            'simple_pose_resnet50_v1d', 'simple_pose_resnet101_v1d',
            'simple_pose_resnet152_v1d']
 
+import gluoncv as gcv
 from mxnet.context import cpu
 from mxnet.gluon.block import HybridBlock
 from mxnet.gluon import nn
@@ -33,7 +34,8 @@ class SimplePoseResNet(HybridBlock):
         super(SimplePoseResNet, self).__init__(**kwargs)
 
         from ..model_zoo import get_model
-        base_network = get_model(base_name, pretrained=pretrained_base, ctx=pretrained_ctx)
+        base_network = get_model(base_name, pretrained=pretrained_base, ctx=pretrained_ctx,
+                                 norm_layer=gcv.nn.BatchNormCudnnOff)
 
         self.resnet = nn.HybridSequential()
         if base_name.endswith('v1'):
@@ -98,8 +100,8 @@ class SimplePoseResNet(HybridBlock):
                         use_bias=self.deconv_with_bias,
                         weight_initializer=initializer.Normal(0.001),
                         bias_initializer=initializer.Zero()))
-                layer.add(nn.BatchNorm(gamma_initializer=initializer.One(),
-                                       beta_initializer=initializer.Zero()))
+                layer.add(gcv.nn.BatchNormCudnnOff(gamma_initializer=initializer.One(),
+                                                   beta_initializer=initializer.Zero()))
                 layer.add(nn.Activation('relu'))
                 self.inplanes = planes
