+++ gluoncv/utils/viz/__init__.py
@@ -3,5 +3,6 @@ from __future__ import absolute_import
 
 from .image import plot_image
 from .bbox import plot_bbox
+from .keypoints import plot_keypoints
 from .mask import expand_mask, plot_mask
 from .segmentation import get_color_pallete, DeNormalize
