+++ scripts/keypoints/simple_pose/coco.sh
