+++ scripts/keypoints/simple_pose/convert_validate.sh
