+++ scripts/keypoints/simple_pose/train_simple_pose.py
