+++ scripts/keypoints/simple_pose/validate.py
