+++ scripts/keypoints/simple_pose/validate.sh
